[feat] create Method: getStockById S11P21A108-54

+++ backend/src/main/java/com/example/donzoom/repository/StockRepository.java
@@ -3,8 +3,9 @@ package com.example.donzoom.repository;
 
 import com.example.donzoom.entity.Stock;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockRepository extends JpaRepository<Stock, Integer> {
+public interface StockRepository extends JpaRepository<Stock, Long> {
 
 }
